requirements.txt
Flask
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
